Move applyBps to bigint utils
The applyBps helper is a generic basis-points operation on a
bigint, not fee-specific — both the slippage and fee code use it.
Move it (and BPS_DENOMINATOR) out of utils/fees into utils/bigint
alongside the other bigint helpers, and relocate its unit tests to
match. Fee-specific helpers (weiToUsd, sumFees) stay in utils/fees.
