+++ web/src/fetchers/fetchTotalRedeemFees.ts
@@ -6,7 +6,8 @@ import { redeemFeeOptions } from "hooks/useRedeemFee";
 import { redeemGasUnitsOptions } from "hooks/useSwapRedeemFees";
 import { config } from "providers/web3Provider";
 import type { TokenWithGateway } from "types";
-import { applyBps, weiToUsd } from "utils/fees";
+import { applyBps } from "utils/bigint";
+import { weiToUsd } from "utils/fees";
 import { getTokenPrice } from "utils/token";
 import { type Address, type Chain, type Client, formatUnits } from "viem";
