+++ web/test/utils/fees.test.ts
@@ -1,27 +1,5 @@
-import { applyBps, sumFees, weiToUsd } from "utils/fees";
+import { sumFees, weiToUsd } from "utils/fees";
 import { describe, expect, it } from "vitest";
-
-describe("applyBps", function () {
-  it("calculates 0.30% of 10000", function () {
-    expect(applyBps(10000n, 30n)).toBe(30n);
-  });
-
-  it("calculates 1% of 1000000", function () {
-    expect(applyBps(1000000n, 100n)).toBe(10000n);
-  });
-
-  it("returns 0 when amount is 0", function () {
-    expect(applyBps(0n, 50n)).toBe(0n);
-  });
-
-  it("returns 0 when bps is 0", function () {
-    expect(applyBps(1000n, 0n)).toBe(0n);
-  });
-
-  it("truncates when result is not whole", function () {
-    expect(applyBps(1n, 5000n)).toBe(0n);
-  });
-});
 
 describe("sumFees", function () {
   it("returns undefined when any fee is undefined", function () {
